requirements.txt
pandas==2.3.1
python-dotenv==1.1.1
requests==2.32.5
streamlit==1.48.1

